Parking sensor with LEDs
This project aims to simulate the behavior of parking sensors in cars. Instead of the standard approach where the closer an object is, the more LED lights illuminate, here we have the opposite relationship.

When an object is close to the sensors, fewer LED lights light up on the display, signaling to the driver that they are approaching an obstacle. Conversely, as the object moves away from the sensors, more LED lights become visible, indicating a greater distance from the obstacle.
In a future update of the code, I will reverse the logic of the lights and add logic responsible for generating sounds like in a real car, as well as mount a speaker on the board.

Watch the lights in action here: https://youtu.be/LkFHq7YeRMg

As a student encountering such a sensor for the first time and taking my first steps on the Arduino platform I utilized the knowledge and code snippets provided on the website: howtomechatronics.com/tutorials/arduino/ultrasonic-sensor-hc-sr04, and I do not claim intellectual ownership of the code fragment responsible for calculating the distance.