Lab 6 - parking sensor with implemented LEDs according to the automotive standard - improved lab 4
This project aims to simulate the behavior of parking sensors in cars. The code follows the standard approach, where the closer an object is, the more LED lights illuminate.

In a future update of the code, I will reverse the logic of the lights and add logic responsible for generating sounds like in a real car, as well as mount a speaker on the board.